+++ docs/extensions/extensions.data.ts
@@ -278,6 +278,32 @@ const baseExtensions: Extension[] = [
     size: 2635,
   },
   {
+    name: 'pgtap',
+    description: `
+    pgTAP is a suite of database functions that make it easy to write TAP-emitting unit 
+    tests in psql scripts or xUnit-style test functions. The TAP output is suitable for 
+    harvesting, analysis, and reporting by a TAP harness, such as those used in Perl 
+    applications.
+    `,
+    shortDescription: 'pgTAP',
+    docs: 'https://pgtap.org',
+    tags: ['postgres extension'],
+    importPath: '@electric-sql/pglite/pgtap',
+    importName: 'pgtap',
+    size: 239428,
+  },
+  {
+    name: 'pg_uuidv7',
+    description: `
+    A tiny Postgres extension to create valid version 7 UUIDs in Postgres.`,
+    shortDescription: 'Use the new v7 UUIDs in Postgres',
+    docs: 'https://github.com/fboulnois/pg_uuidv7/',
+    tags: ['postgres extension'],
+    importPath: '@electric-sql/pglite/pg_uuidv7',
+    importName: 'pg_uuidv7',
+    size: 1522,
+  },
+  {
     name: 'pg_walinspect',
     description: `
       The pg_walinspect module provides SQL functions that allow you to inspect the