+++ packages/pglite-tools/src/pg_dump.ts
@@ -1,69 +1,28 @@
-import { WasiPreview1 } from './wasi/easywasi'
-import { postgresMod } from '@electric-sql/pglite'
 import { PGlite } from '@electric-sql/pglite'
+import PgDumpModFactory, { PgDumpMod } from './pgDumpModFactory'
 
-type FS = postgresMod.FS
-type FSInterface = any // WASI FS interface
-
-const IN_NODE =
-  typeof process === 'object' &&
-  typeof process.versions === 'object' &&
-  typeof process.versions.node === 'string'
+const dumpFilePath = '/tmp/out.sql'
 
 /**
- * Emscripten FS is not quite compatible with WASI
- * so we need to patch it
+ * Creates a new Uint8Array based on two different ArrayBuffers
+ *
+ * @private
+ * @param {ArrayBuffers} buffer1 The first buffer.
+ * @param {ArrayBuffers} buffer2 The second buffer.
+ * @return {ArrayBuffers} The new ArrayBuffer created out of the two.
  */
-function emscriptenFsToWasiFS(fs: FS, acc: any[]): FSInterface & FS {
-  const requiredMethods = [
-    'appendFileSync',
-    'fsyncSync',
-    'linkSync',
-    'setFlagsSync',
-    'mkdirSync',
-    'readdirSync',
-    'readFileSync',
-    'readlinkSync',
-    'renameSync',
-    'rmdirSync',
-    'statSync',
-    'symlinkSync',
-    'truncateSync',
-    'unlinkSync',
-    'utimesSync',
-    'writeFileSync',
-  ]
-  return {
-    // Bind all methods to the FS instance
-    ...fs,
-    // Add missing methods
-    ...Object.fromEntries(
-      requiredMethods
-        .map((method) => {
-          const target = method.slice(0, method.length - 4)
-          if (!(target in fs)) {
-            return [
-              method,
-              () => {
-                throw new Error(`${target} not implemented.`)
-              },
-            ]
-          }
-          return [
-            method,
-            (...args: any[]) => {
-              if (method === 'writeFileSync' && args[0] === '/tmp/out.sql') {
-                acc.push(args[1])
-              }
-              return (fs as any)[target](...args)
-            },
-          ]
-        })
-        .filter(
-          (entry): entry is [string, (fs: FS) => any] => entry !== undefined,
-        ),
-    ),
-  } as FSInterface & FS
+function concat(buffer1: ArrayBuffer, buffer2: ArrayBuffer) {
+  const tmp = new Uint8Array(buffer1.byteLength + buffer2.byteLength)
+  tmp.set(new Uint8Array(buffer1), 0)
+  tmp.set(new Uint8Array(buffer2), buffer1.byteLength)
+  return tmp
+}
+
+interface ExecResult {
+  exitCode: number
+  fileContents: string
+  stderr: string
+  stdout: string
 }
 
 /**
@@ -75,87 +34,82 @@ async function execPgDump({
 }: {
   pg: PGlite
   args: string[]
-}): Promise<[number, Uint8Array[], string]> {
-  const bin = new URL('./pg_dump.wasm', import.meta.url)
-  const acc: Uint8Array[] = []
-  const FS = emscriptenFsToWasiFS(pg.Module.FS, acc)
-
-  const wasi = new WasiPreview1({
-    fs: FS,
-    args: ['pg_dump', ...args],
-    env: {
-      PWD: '/',
+}): Promise<ExecResult> {
+  let pgdump_write, pgdump_read
+  let exitStatus = 0
+  let stderrOutput: string = ''
+  let stdoutOutput: string = ''
+  const emscriptenOpts: Partial<PgDumpMod> = {
+    arguments: args,
+    noExitRuntime: false,
+    print: (text) => {
+      stdoutOutput += text
     },
-  })
+    printErr: (text) => {
+      stderrOutput += text
+    },
+    onExit: (status: number) => {
+      exitStatus = status
+    },
+    preRun: [
+      (mod: PgDumpMod) => {
+        mod.onRuntimeInitialized = () => {
+          let bufferedBytes: Uint8Array = new Uint8Array()
 
-  wasi.stdout = (_buffer) => {
-    // console.log('stdout', _buffer)
-  }
-  const textDecoder = new TextDecoder()
-  let errorMessage = ''
+          pgdump_write = mod.addFunction((ptr: any, length: number) => {
+            let bytes
+            try {
+              bytes = mod.HEAPU8.subarray(ptr, ptr + length)
+            } catch (e: any) {
+              console.error('error', e)
+              throw e
+            }
+            const currentResponse = pg.execProtocolRawSync(bytes)
+            bufferedBytes = concat(bufferedBytes, currentResponse)
+            return length
+          }, 'iii')
 
-  wasi.stderr = (_buffer) => {
-    const text = textDecoder.decode(_buffer)
-    if (text) errorMessage += text
-  }
-  wasi.sched_yield = () => {
-    const pgIn = '/tmp/pglite/base/.s.PGSQL.5432.in'
-    const pgOut = '/tmp/pglite/base/.s.PGSQL.5432.out'
-    if (FS.analyzePath(pgIn).exists) {
-      // call interactive one
-      const msgIn = FS.readFileSync(pgIn)
+          pgdump_read = mod.addFunction((ptr: any, max_length: number) => {
+            let length = bufferedBytes.length
+            if (length > max_length) {
+              length = max_length
+            }
+            try {
+              mod.HEAP8.set(bufferedBytes.subarray(0, length), ptr)
+            } catch (e) {
+              console.error(e)
+            }
+            bufferedBytes = bufferedBytes.subarray(length, bufferedBytes.length)
+            return length
+          }, 'iii')
 
-      // BYPASS the file socket emulation in PGlite
-      FS.unlinkSync(pgIn)
-
-      // Handle auth request
-      if (msgIn[0] === 0) {
-        const reply = new Uint8Array([
-          ...authOk,
-          ...versionParam,
-          ...readyForQuery,
-        ])
-        FS.writeFileSync(pgOut, reply)
-        return 0
-      }
-
-      // Handle query
-      const reply = pg.execProtocolRawSync(msgIn)
-      FS.writeFileSync(pgOut, reply)
-    }
-    return 0
-  }
-
-  // Postgres can complain if the binary is not on the filesystem
-  // so we create a dummy file
-  await FS.writeFile('/pg_dump', '\0', { mode: 18 })
-
-  let app: WebAssembly.WebAssemblyInstantiatedSource
-
-  if (IN_NODE) {
-    const fs = await import('fs/promises')
-    const blob = await fs.readFile(bin)
-    app = await WebAssembly.instantiate(blob, {
-      wasi_snapshot_preview1: wasi as any,
-    })
-  } else {
-    app = await WebAssembly.instantiateStreaming(fetch(bin), {
-      wasi_snapshot_preview1: wasi as any,
-    })
+          mod._set_read_write_cbs(pgdump_read, pgdump_write)
+          // default $HOME in emscripten is /home/web_user
+          mod.FS.chmod('/home/web_user/.pgpass', 0o0600) // https://www.postgresql.org/docs/current/libpq-pgpass.html
+        }
+      },
+    ],
   }
 
-  let exitCode: number
-  await pg.runExclusive(async () => {
-    exitCode = wasi.start(app.instance.exports)
-  })
+  const mod = await PgDumpModFactory(emscriptenOpts)
+  let fileContents = ''
+  if (!exitStatus) {
+    fileContents = mod.FS.readFile(dumpFilePath, { encoding: 'utf8' })
+  }
 
-  return [exitCode!, acc, errorMessage]
+  return {
+    exitCode: exitStatus,
+    fileContents,
+    stderr: stderrOutput,
+    stdout: stdoutOutput,
+  }
 }
 
 interface PgDumpOptions {
   pg: PGlite
   args?: string[]
   fileName?: string
+  verbose?: boolean
 }
 
 /**
@@ -171,7 +125,6 @@ export async function pgDump({
   )
   const search_path = getSearchPath.rows[0].search_path
 
-  const outFile = `/tmp/out.sql`
   const baseArgs = [
     '-U',
     'postgres',
@@ -179,69 +132,26 @@ export async function pgDump({
     '-j',
     '1',
     '-f',
-    outFile,
+    dumpFilePath,
     'postgres',
   ]
 
-  const [exitCode, acc, errorMessage] = await execPgDump({
+  const execResult = await execPgDump({
     pg,
     args: [...(args ?? []), ...baseArgs],
   })
 
   pg.exec(`DEALLOCATE ALL; SET SEARCH_PATH = ${search_path}`)
 
-  if (exitCode !== 0) {
+  if (execResult.exitCode !== 0) {
     throw new Error(
-      `pg_dump failed with exit code ${exitCode}. \nError message: ${errorMessage}`,
+      `pg_dump failed with exit code ${execResult.exitCode}. \nError message: ${execResult.stderr}`,
     )
   }
 
-  const file = new File(acc, fileName, {
+  const file = new File([execResult.fileContents], fileName, {
     type: 'text/plain',
   })
-  pg.Module.FS.unlink(outFile)
 
   return file
 }
-
-// Wire protocol messages for simulating auth handshake:
-
-function charToByte(char: string) {
-  return char.charCodeAt(0)
-}
-
-// Function to convert an integer to a 4-byte array (Int32)
-function int32ToBytes(value: number) {
-  const buffer = new ArrayBuffer(4)
-  const view = new DataView(buffer)
-  view.setInt32(0, value, false) // false for big-endian
-  return new Uint8Array(buffer)
-}
-
-// Convert a string to a Uint8Array with a null terminator (C string)
-function stringToBytes(str: string) {
-  const utf8Encoder = new TextEncoder()
-  const strBytes = utf8Encoder.encode(str) // UTF-8 encoding
-  return new Uint8Array([...strBytes, 0]) // Append null terminator
-}
-
-const authOk = new Uint8Array([
-  charToByte('R'),
-  ...int32ToBytes(8),
-  ...int32ToBytes(0),
-])
-const readyForQuery = new Uint8Array([
-  charToByte('Z'),
-  ...int32ToBytes(5),
-  charToByte('I'),
-])
-
-const svParamName = stringToBytes('server_version')
-const svParamValue = stringToBytes('16.3 (PGlite 0.2.0)')
-const svTotalLength = 4 + svParamName.length + svParamValue.length
-const versionParam = new Uint8Array([
-  charToByte('S'),
-  ...int32ToBytes(svTotalLength),
-  ...svParamName,
-  ...svParamValue,
-])