+++ packages/pglite-repl/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @electric-sql/pglite-repl
 
+## 0.2.31
+
+### Patch Changes
+
+- Updated dependencies [ad3d0d8]
+  - @electric-sql/pglite@0.3.13
+  - @electric-sql/pglite-react@0.2.31
+
+## 0.2.30
+
+### Patch Changes
+
+- Updated dependencies [ce0e74e]
+  - @electric-sql/pglite@0.3.12
+  - @electric-sql/pglite-react@0.2.30
+
 ## 0.2.29
 
 ### Patch Changes