+++ packages/pglite-vue/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @electric-sql/pglite-vue
 
+## 0.2.31
+
+### Patch Changes
+
+- Updated dependencies [ad3d0d8]
+  - @electric-sql/pglite@0.3.13
+
+## 0.2.30
+
+### Patch Changes
+
+- Updated dependencies [ce0e74e]
+  - @electric-sql/pglite@0.3.12
+
 ## 0.2.29
 
 ### Patch Changes